feat: create modal comp with style on click of del

+++ src/components/invoice/InvoiceDetails.vue
@@ -1,25 +1,46 @@
 <template>
-  <section class="details">
-    <app-back-nav></app-back-nav>
-    <div class="details__actions">
-      <div>
-        <p>Status</p>
-        <app-status-pill :stat="invoice.status"></app-status-pill>
+  <div>
+    <section class="details">
+      <app-back-nav></app-back-nav>
+      <div class="details__actions">
+        <div>
+          <p>Status</p>
+          <app-status-pill :stat="invoice.status"></app-status-pill>
+        </div>
+        <div>
+          <app-button type="edit">
+            <p>Edit</p>
+          </app-button>
+          <app-button type="delete" @click.native="showModal = true">
+            <p>Delete</p>
+          </app-button>
+          <app-button type="default">
+            <p>Mark as Paid</p>
+          </app-button>
+        </div>
       </div>
-      <div>
-        <app-button type="edit">
-          <p>Edit</p>
-        </app-button>
-        <app-button type="delete">
-          <p>Delete</p>
-        </app-button>
-        <app-button type="default">
-          <p>Mark as Paid</p>
-        </app-button>
+      <app-invoice-info :invoice="invoice"></app-invoice-info>
+    </section>
+    <app-modal v-if="showModal" @close="showModal = false">
+      <div class="delete">
+        <div class="delete__text">
+          <h2>Confirm Deletion</h2>
+          <p class="p-two">
+            Are you sure you want to delete <span>&#35;</span>{{ invoice.id }}?
+            This action cannot be undone.
+          </p>
+        </div>
+        <div class="delete__btns">
+          <app-button type="edit" @click.native="showModal = false">
+            <p>Cancel</p>
+          </app-button>
+          <app-button type="delete">
+            <p>Delete</p>
+          </app-button>
+        </div>
       </div>
-    </div>
-    <app-invoice-info :invoice="invoice"></app-invoice-info>
-  </section>
+    </app-modal>
+  </div>
 </template>
 
 <script>
@@ -28,6 +49,7 @@ import BackNav from "@/components/shared/BackNav";
 import StatusPill from "@/components/shared/StatusPill";
 import ActionButton from "@/components/shared/ActionButton";
 import InvoiceInfo from "@/components/invoice/InvoiceInfo";
+import AppModal from "@/components/widgets/AppModal";
 export default {
   name: "invoiceDetails",
   computed: mapState(["invoice"]),
@@ -36,6 +58,12 @@ export default {
     "app-status-pill": StatusPill,
     "app-button": ActionButton,
     "app-invoice-info": InvoiceInfo,
+    "app-modal": AppModal,
+  },
+  data() {
+    return {
+      showModal: false,
+    };
   },
 };
 </script>
@@ -63,6 +91,24 @@ export default {
       align-items: center;
       gap: 1rem;
     }
+  }
+}
+
+.delete {
+  &__text {
+    margin-bottom: 2rem;
+    h2 {
+      margin-bottom: 1.5rem;
+    }
+    span {
+      font-size: 1.1rem;
+      color: var(--clr-grey-two);
+    }
+  }
+  &__btns {
+    display: flex;
+    justify-content: flex-end;
+    gap: 1rem;
   }
 }
 </style>
